don't clone repo if it already exists

+++ entrypoint.sh
@@ -18,9 +18,11 @@ case "${1:-api}" in
         exec httpd -DFOREGROUND
         ;;
     "daemon")
-        echo Cloning $REPO_NAME from $REPO_URL
-        hg clone $REPO_URL /repos/$REPO_NAME
-        envsubst < /hgrc.template > /repos/$REPO_NAME/.hg/hgrc
+        if [[ ! -e /repos/$REPO_NAME ]]; then
+            echo Cloning $REPO_NAME from $REPO_URL
+            hg clone $REPO_URL /repos/$REPO_NAME
+            envsubst < /hgrc.template > /repos/$REPO_NAME/.hg/hgrc
+        fi
         cd $AUTOLAND_HOME
         . venv/bin/activate
         cd autoland
